refactor(app): shift reorganize file on shift module
WIP
