+++ app/(dashboard)/dashboard/organization/branches/table/data-table.tsx
@@ -22,72 +22,6 @@ interface DataTableProps<TData, TValue> {
     data: TData[];
 }
 
-// export function DataTable<TData, TValue>({
-//     columns,
-//     data,
-// }: DataTableProps<TData, TValue>) {
-//     const table = useReactTable({
-//         data,
-//         columns,
-//         getCoreRowModel: getCoreRowModel(),
-//     });
-
-//     return (
-//         <div>
-//             <Table>
-//                 <TableHeader>
-//                     {table.getHeaderGroups().map((headerGroup) => (
-//                         <TableRow key={headerGroup.id}>
-//                             {headerGroup.headers.map((header) => {
-//                                 return (
-//                                     <TableHead key={header.id}>
-//                                         {header.isPlaceholder
-//                                             ? null
-//                                             : flexRender(
-//                                                   header.column.columnDef
-//                                                       .header,
-//                                                   header.getContext()
-//                                               )}
-//                                     </TableHead>
-//                                 );
-//                             })}
-//                         </TableRow>
-//                     ))}
-//                 </TableHeader>
-
-//                 <TableBody>
-//                     {table.getRowModel().rows?.length ? (
-//                         table.getRowModel().rows.map((row) => (
-//                             <TableRow
-//                                 key={row.id}
-//                                 data-state={row.getIsSelected() && "selected"}
-//                             >
-//                                 {row.getVisibleCells().map((cell) => (
-//                                     <TableCell key={cell.id}>
-//                                         {flexRender(
-//                                             cell.column.columnDef.cell,
-//                                             cell.getContext()
-//                                         )}
-//                                     </TableCell>
-//                                 ))}
-//                             </TableRow>
-//                         ))
-//                     ) : (
-//                         <TableRow>
-//                             <TableCell
-//                                 colSpan={columns.length}
-//                                 className="h-24 text-center"
-//                             >
-//                                 No results.
-//                             </TableCell>
-//                         </TableRow>
-//                     )}
-//                 </TableBody>
-//             </Table>
-//         </div>
-//     );
-// }
-
 interface DataTableProps<TData, TValue> {
     columns: ColumnDef<TData, TValue>[];
     data: TData[];
