+++ app/(dashboard)/dashboard/setting/shift/page.tsx
@@ -1,16 +1,18 @@
-import { Shift, columns } from "./columns";
-import { DataTable } from "./data-table";
+import { Shift } from "./table/columns";
+import { columns } from "./table/columns";
 import { get } from "@/lib/fetch-wrapper";
 import { Suspense } from "react";
 import Loading from "@/app/loading";
 import CreateShiftModal from "./create-shift-modal";
 import { DataTableC } from "@/components/table/data-table";
+import { DataTable } from "./table/data-table";
 import { DataTableToolbar } from "./data-table-toolbar";
 
+
 async function getShifts(): Promise<Shift[]> {
-    const response = await get("shifts");
+    const response = (await get("shifts", ["shifts"])) as { data: Shift[] };
     // console.log("response :", response);
-    return response as Shift[];
+    return response.data;
 }
 
 export default async function DataTab() {
